html: simply call out to golang.org/x/net/html
It is silly to maintain two identical versions of the EscapeString and
UnescapeString functions.  Over the last few years, they've drifted apart,
and each gotten separate fixes and improvements.  So let's consolidate.

Let x/net have the canonical version and std call it, rather than
vice-versa, because if we had x/net call std, we'd need to expand the API
to have []byte versions, and to be able to pass the `attribute` argument.